+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -5,30 +5,6 @@ from typing import Dict, Any, Optional
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
     @mcp.tool()
-    def wafw00f_scan(target: str, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute wafw00f to identify and fingerprint WAF products with enhanced logging.
-
-        Args:
-            target: Target URL or IP
-            additional_args: Additional wafw00f arguments
-
-        Returns:
-            WAF detection results
-        """
-        data = {
-            "target": target,
-            "additional_args": additional_args
-        }
-        logger.info(f"🛡️ Starting Wafw00f WAF detection: {target}")
-        result = hexstrike_client.safe_post("api/tools/wafw00f", data)
-        if result.get("success"):
-            logger.info(f"✅ Wafw00f completed for {target}")
-        else:
-            logger.error(f"❌ Wafw00f failed for {target}")
-        return result
-
-    @mcp.tool()
     def fierce_scan(domain: str, dns_server: str = "", additional_args: str = "") -> Dict[str, Any]:
         """
         Execute fierce for DNS reconnaissance with enhanced logging.
